Set state only when return value is -1
It took me way too long to wrap my head around this, but if the result of `Array.findIndex` is greater than -1, that means the element already exists and we should not be logging it. Instead, we need to check if the return value is -1 (no index found). It also returns -1 if the array is empty, so it also works when the page is initially loaded