refactor: modernize code and remove dependencies (#260)
This PR refactors the whole codebase. Removes unecessary deps, use of
built in fetch.

Should be released as semver-major.

I was first confused, because it seemed, that it was bugged, because the
payload of the endpoint was not returned. Then I tested it with the
original smee-client and realized that it is a one way thing... smee.io
does not return the answer of the webserver. Please correct me if I am
wrong.

Coverage is improved significantly.

```sh
 % Coverage report from v8
----------|---------|----------|---------|---------|-------------------
File      | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s 
----------|---------|----------|---------|---------|-------------------
All files |   94.49 |    88.88 |   66.66 |   94.49 |                   
 index.ts |   94.49 |    88.88 |   66.66 |   94.49 | 80-81,85-86,89-90 
----------|---------|----------|---------|---------|-------------------
```

BREAKING CHANGE: drop support for Node <18

BREAKING CHANGE: use native `fetch` for requests